created initial typescript setup
- migrated rollup configuration to typescript
- removed index.d.ts, since we now emit declarations with tsc
- removed dev-dependency `rollup-plugin-dts`
